app/requirements.txt
amqp==5.0.6
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
decorator==5.0.9
Django==3.2.4
django-cors-headers==3.7.0
django-settings-export==1.2.1
djangorestframework==3.12.4
importlib-metadata==4.5.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.1.0
parso==0.8.2
pdfkit==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.1